remove SimulationEngine.py (moved into engines)
